README.*: Add `sudo(1)` to example commands
The previous instructions implied that installing with privileges is as
simple as turning `make install` into `sudo make install`.  The problem
there is that the *whole build* runs with privileges and may populate
the build directory with files that can't be updated or removed without
them.

Avoid possible confusion by revising the example installation commands
to use the idiomatic `make && sudo make install`, which uses privileges
for the `install` target only.  Add `sudo` to the `uninstall` examples
too, for consistency.