build: Rename some macros
Rename `cleanup` to `do_cleanup` to suggest its purpose as a standalone
"command".

Rename `progs` to `bin_SCRIPTS` because the latter is a special Automake
variable, so it may be useful to readers who are familiar with Automake.
(For the rest of us, neither name is obviously better than the other.)